Fix length var for preload
got NumberFormatException from Int in TorrentHelper/waitFiles

+++ app/src/main/java/ru/yourok/torrserve/server/models/torrent/Torrent.kt
@@ -37,5 +37,5 @@ data class Torrent(
 data class FileStat(
     var id: Int,
     var path: String,
-    var length: Int
+    var length: Long
 )
